PFP-9002 som nevnt i saken flyttet behandlingPåVent,venteÅrsak og fristBehandlingPåVent fra UtvidetBehandlingDto til BehandlingDto
